+++ src/test/java/uk/gov/hmcts/divorce/citizen/event/CitizenApplicant2UpdateApplicationTest.java
@@ -5,13 +5,18 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.ccd.sdk.ConfigBuilderImpl;
+import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
 import uk.gov.hmcts.ccd.sdk.api.Event;
+import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
+import uk.gov.hmcts.ccd.sdk.type.YesOrNo;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.divorcecase.model.HowToRespondApplication;
 import uk.gov.hmcts.divorce.divorcecase.model.State;
 import uk.gov.hmcts.divorce.divorcecase.model.UserRole;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static uk.gov.hmcts.divorce.citizen.event.CitizenApplicant2UpdateApplication.CITIZEN_APPLICANT2_UPDATE;
+import static uk.gov.hmcts.divorce.divorcecase.model.State.AosDrafted;
 import static uk.gov.hmcts.divorce.testutil.ConfigTestUtil.createCaseDataConfigBuilder;
 import static uk.gov.hmcts.divorce.testutil.ConfigTestUtil.getEventsFrom;
 
@@ -30,5 +35,24 @@ public class CitizenApplicant2UpdateApplicationTest {
         assertThat(getEventsFrom(configBuilder).values())
             .extracting(Event::getId)
             .contains(CITIZEN_APPLICANT2_UPDATE);
+    }
+
+    @Test
+    void shouldSetDisputeApplicationFieldsToNullIfConfirmationIsNo() {
+        final long caseId = 1L;
+        final CaseDetails<CaseData, State> caseDetails = new CaseDetails<>();
+        final CaseData caseData = CaseData.builder().build();
+        caseDetails.setState(AosDrafted);
+        caseData.getAcknowledgementOfService().setHowToRespondApplication(HowToRespondApplication.DISPUTE_DIVORCE);
+        caseData.getAcknowledgementOfService().setConfirmDisputeApplication(YesOrNo.NO);
+
+        caseDetails.setData(caseData);
+        caseDetails.setId(caseId);
+
+        final AboutToStartOrSubmitResponse<CaseData, State> response
+            = citizenApplicant2UpdateApplication.aboutToSubmit(caseDetails, caseDetails);
+
+        assertThat(response.getData().getAcknowledgementOfService().getHowToRespondApplication()).isNull();
+        assertThat(response.getData().getAcknowledgementOfService().getConfirmDisputeApplication()).isNull();
     }
 }
