+++ src/test/java/uk/gov/hmcts/divorce/common/event/UpdateAosTest.java
@@ -9,9 +9,7 @@ import uk.gov.hmcts.ccd.sdk.ConfigBuilderImpl;
 import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
 import uk.gov.hmcts.ccd.sdk.api.Event;
 import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
-import uk.gov.hmcts.ccd.sdk.type.YesOrNo;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
-import uk.gov.hmcts.divorce.divorcecase.model.HowToRespondApplication;
 import uk.gov.hmcts.divorce.divorcecase.model.State;
 import uk.gov.hmcts.divorce.divorcecase.model.UserRole;
 import uk.gov.hmcts.divorce.solicitor.service.task.AddMiniApplicationLink;
@@ -58,24 +56,5 @@ class UpdateAosTest {
         assertThat(response.getData()).isSameAs(expectedCaseData);
 
         verify(addMiniApplicationLink).apply(caseDetails);
-    }
-
-
-    @Test
-    void shouldSetDisputeApplicationFieldsToNullIfConfirmationIsNo() {
-        final long caseId = 1L;
-        final CaseDetails<CaseData, State> caseDetails = new CaseDetails<>();
-        final CaseData caseData = CaseData.builder().build();
-        caseData.getAcknowledgementOfService().setHowToRespondApplication(HowToRespondApplication.DISPUTE_DIVORCE);
-        caseData.getAcknowledgementOfService().setConfirmDisputeApplication(YesOrNo.NO);
-
-        caseDetails.setData(caseData);
-        caseDetails.setId(caseId);
-
-        final AboutToStartOrSubmitResponse<CaseData, State> response = updateAos.aboutToSubmit(caseDetails, caseDetails);
-
-        assertThat(response.getData().getAcknowledgementOfService().getHowToRespondApplication()).isNull();
-        assertThat(response.getData().getAcknowledgementOfService().getConfirmDisputeApplication()).isNull();
-
     }
 }
